+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added function `Format-HRRoster` to public functions to format roster for compare.
+
 ## [1.6.1] - 2022-07-21
 
 ### Added